Moved original client to clients subdir. Moved place_requests into simple_client and renamed file to wait_requests
