+++ include/client/client_wait_requests.h
@@ -1,5 +1,5 @@
-#ifndef CLIENT_PLACE_REQUESTS
-#define CLIENT_PLACE_REQUESTS
+#ifndef CLIENT_WAIT_REQUESTS
+#define CLIENT_WAIT_REQUESTS
 
 #include "client.h"
 #include "request.h"
@@ -7,6 +7,5 @@
 void wait_request(struct client *client, struct request *req);
 void wait_requests(struct client *client, struct request *reqs,
                    size_t num_reqs);
-void place_requests(struct client *client);
 
-#endif /* CLIENT_PLACE_REQUESTS */
+#endif /* CLIENT_WAIT_REQUESTS */
