+++ src/client/CMakeLists.txt
@@ -9,6 +9,6 @@ set(CLIENT_SOURCE_FILES ${CLIENT_SOURCE_FILES}
                         ${CMAKE_CURRENT_SOURCE_DIR}/client_cleanup.c 
                         ${CMAKE_CURRENT_SOURCE_DIR}/client_init.c 
                         ${CMAKE_CURRENT_SOURCE_DIR}/client_memory.c 
-                        ${CMAKE_CURRENT_SOURCE_DIR}/client_place_requests.c 
+                        ${CMAKE_CURRENT_SOURCE_DIR}/client_wait_requests.c 
                         ${CMAKE_CURRENT_SOURCE_DIR}/client_read_location.c 
                         PARENT_SCOPE)
