+++ src/client/client_wait_requests.c
@@ -1,4 +1,4 @@
-#include "client_place_requests.h"
+#include "client_wait_requests.h"
 #include "client_memory.h"
 
 #include "client.h"
@@ -61,47 +61,3 @@ void wait_requests(struct client *client, struct request *reqs,
     wait_request(client, &reqs[i]);
 }
 
-void place_requests(struct client *client) {
-  // Allocate Buffer
-  double *A = shm_malloc(1024 * sizeof(int));
-
-  if (client->chatty)
-    printf("Client(%d): Received Pointer to Allocated Memory: %p\n", client->id,
-           (void *)A);
-
-  // Write to Buffer
-  printf("Client(%d): Writing Array:\n", client->id);
-
-  double *input = shm_malloc(1024 * sizeof(double));
-  for (size_t i = 0; i < 1024; ++i) {
-    input[i] = (double)(2 * i);
-  }
-
-  struct request req1;
-  scoria_write(client, A, 1024, input, NULL, NULL, 0, 0, &req1);
-  wait_request(client, &req1);
-  shm_free(input);
-
-  // Read from Buffer
-  printf("Client(%d): Reading Array:\n", client->id);
-
-  double *output = shm_malloc(1024 * sizeof(double));
-
-  struct request req2;
-  scoria_read(client, A, 1024, output, NULL, NULL, 0, 0, &req2);
-  wait_request(client, &req2);
-
-  for (size_t i = 0; i < 1024; ++i)
-    printf("%.2f ", output[i]);
-  printf("\n");
-  shm_free(output);
-
-  // Free Buffer
-  shm_free(A);
-
-  // Exit Program
-  // struct request req3;
-  // scoria_quit(client, &req3);
-
-  // wait_request(client, &req3);
-}
